requirements-handler.txt
# Requirements for RunPod Text-to-Image Handler
runpod>=1.7.0
requests>=2.28.0
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
# Compatible with base image PyTorch 2.2.1
torchvision==0.17.1
xformers>=0.0.24
boto3>=1.34.0
